requirements.txt
boto3==1.7.20
botocore==1.10.20
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.5
django-appconf==1.0.2
django-imagekit==4.0.2
django-inline-svg==0.1.1
django-leaflet==0.23.0
django-localflavor==2.0
django-nested-admin==3.0.21
django-sortedm2m==1.5.0
django-storages==1.6.6
docutils==0.14
future==0.16.0
idna==2.6
jmespath==0.9.3
mapbox-vector-tile==1.2.0
Markdown==2.6.11
martor==1.2.5
mercantile==1.0.2
pbr==4.0.2
pilkit==2.0
Pillow==5.1.0
protobuf==3.5.2.post1
pyclipper==1.1.0
python-dateutil==2.7.3
python-monkey-business==1.0.0
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
selenium==3.12.0
Shapely==1.6.4.post1
six==1.11.0
urllib3==1.22
